refactor: 抽取 Emby API 共用函数到 emby_shared.py
将 _generate_server_url/_build_jellyfin_headers/_is_jellyfin_server/
_append_query/_upload_actor_photo 五个完全一致的函数从 emby_actor_image
和 emby_actor_manager 中提取到新建的 emby_shared 模块。

两个模块均从 emby_shared 导入并 re-export，保持向后兼容。
消除了 manager 中两处 _upload_actor_photo 的 lazy import。
